+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # EnhancedJDA - Changelog
 
+## 1.0.2 - 04.01.2025
+
+- Controller won't try to connect with database when it's useless
+- Added check for unknown database type
+- Added info about registered persisters and tables
+- Added option to pass EnhancedBot into persister constructor
+- Added guild persister
+
 ## 1.0.1 - 02.01.2025
 
 - Improved ReflectionUtil & Controllers to be able to test bot by running it using IDE